fix: Support TypeScript module imports (#7)
* fix: Support TypeScript module imports

Adds support for importing TypeScript modules (`.ts` and `.tsx` files) in the manifest, enabling their inclusion in the build process.

* Fix: Prepend resource paths with `/`

Ensures that resource URLs are treated as root-relative paths. This resolves an issue where resources were being fetched from incorrect locations, especially when the application was deployed on a non-root path.